test(backup): Improve backup tests
This change contains a number of small improvements, including:

- Renaming some files for greater clarity
- Generalizing the `BackupTestCase` class, which creates an
  "exhaustive" database state for testing backups. We'll be using this
  much more in the future as we implement features like org-based
  filtering.
- Sped up `test_models` and `test_releases` considerably. These tests
  both inherit from `TransactionTestCase`, which already does the heavy
  lifting of rolling back any DB changes made during a test. However, we
  were also manually flushing the database during the `setUp` phase,
  resuling in a repeat of this super time consuming action. Removing it
  makes the average test case in one of these files go from ~6 seconds
  to less than 3!
- Because we are now running various tests with the `reset_pks` flag set
  to `True` or `False` depending on the behavior we want to exercise,
  I've added a check to `import_export_then_validate` to ensure that pks
  have either all changes or all stayed the same, depending on the flag
  setting.

Issue: getsentry/team-ospo#167
