+++ tests/sentry/backup/test_coverage.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from sentry.backup.helpers import get_exportable_final_derivations_of
 from sentry.db.models import BaseModel
 from tests.sentry.backup.test_models import UNIT_TESTED_MODELS
-from tests.sentry.backup.test_releases import RELEASE_TESTED_MODELS
+from tests.sentry.backup.test_release import RELEASE_TESTED_MODELS
 
 ALL_EXPORTABLE_MODELS = {c.__name__ for c in get_exportable_final_derivations_of(BaseModel)}
